Revert "revamp verification and add expiration"
This reverts commit 175582989a856acd30bcd6695e295c58296a226a.
