--- cmd/verify/app/main.go
@@ -1,21 +1,0 @@
-package app
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/spf13/cobra"
-)
-
-var rootCmd = &cobra.Command{
-	Use:   "verify",
-	Short: "Root-signing verify CLI",
-	Long:  `Root-signing verify command line interface tool`,
-}
-
-func Execute() {
-	if err := rootCmd.Execute(); err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-	}
-}
